requirements.txt
Django>=2.2.5
djangorestframework==3.11.0
psycopg2==2.8.4
django-filter==2.2.0